refactor(PawnMoveStrategy) : 폰 기물의 중복 로직을 추상화
