+++ src/main/java/chess/domain/strategy/BlackPawnMoveStrategy.java
@@ -1,37 +1,14 @@
 package chess.domain.strategy;
 
 import chess.domain.Direction;
-import chess.domain.Position;
 import chess.domain.Row;
-import java.util.ArrayDeque;
-import java.util.Deque;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
-public class BlackPawnMoveStrategy implements MoveStrategy{
+public class BlackPawnMoveStrategy extends PawnMoveStrategy{
 
-    private static final int DEFAULT_MAX_MOVE_DISTANCE = 1;
+    private static final List<Direction> DIRECTIONS = List.of(Direction.S, Direction.SE, Direction.SW);
 
-    private final List<Direction> directions = List.of(Direction.S, Direction.SE, Direction.SW);
-
-    @Override
-    public Map<Direction, Deque<Position>> generateMovablePositions(Position position) {
-        return directions.stream()
-                .collect(Collectors.toMap(
-                        direction -> direction,
-                        direction -> generateMovablePositionsByDirection(position, direction)
-                ));
-    }
-
-    private Deque<Position> generateMovablePositionsByDirection(Position currentPosition, Direction direction) {
-        int movableMaxDistance = currentPosition.calculateMaxDistance(direction, DEFAULT_MAX_MOVE_DISTANCE);
-        if (currentPosition.isSameRow(Row.RANK7) && direction == Direction.S) {
-            movableMaxDistance++;
-        }
-        return new ArrayDeque<>(IntStream.rangeClosed(1, movableMaxDistance)
-                .mapToObj(weight -> currentPosition.calculateNextPosition(direction, weight))
-                .toList());
+    public BlackPawnMoveStrategy() {
+        super(DIRECTIONS, Row.RANK7, Direction.S);
     }
 }
