+++ src/test/java/chess/domain/strategy/BlackPawnMoveStrategyTest.java
@@ -1,7 +1,7 @@
 package chess.domain.strategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 import chess.domain.Column;
 import chess.domain.Direction;
